SCP ML Foundation/GRIR/requirements.txt
grpcio
tensorflow==1.3.0
tensorflow-serving-api==1.3.0
requests
flask
flask-socketio
eventlet
scikit-image
numpy